renamed io folders to config
