+++ src/main/java/org/braekpo1nt/mctmanager/games/mecha/config/WeightedNamespacedKey.java
@@ -1,4 +1,4 @@
-package org.braekpo1nt.mctmanager.games.mecha.io;
+package org.braekpo1nt.mctmanager.games.mecha.config;
 
 public record WeightedNamespacedKey(String namespace, String key, int weight) {
 }
